onboarding modification to instances list/website
added Getting Started header and changed alignment to left. notable that the preview seems to show a green line to the left of it? but i believe that is a visual aid used by github to show that that section changed, and not something i somehow accidentally added. if it is by mistake, i can change it once told how